+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # 1.x branch
 ## 1.1 branch
+### 1.1.3
+* added `BodyParser` class.
+
+### 1.1.2
+* added `$removeWWW` optional parameter to `clean_url()` global function.
+
+### 1.1.1
+* added `first_value()` and `last_value()` global functions.
+
 ### 1.1.0
 * added `assertContainsInstanceOf()` assertion method. Removed
     `assertInstanceOf()` (you can use the original method).